refactor: replace page components to own folders
