+++ react-components/src/components/pages/MainPage/MainPage.tsx
